Добавлены интерсепторы. LoginForm и RegisterForm разделены на два компонента. Добавлен сервис для api запросов.
