+++ src/components/QRCode.vue
