+++ src/router/index.ts
@@ -1,5 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import RegisterOrLogin from '@/components/Pages/RegisterOrLogin.vue'
+import RegisterOrLogin from '@/components/Pages/Auth.vue'
+import Auth from '@/components/Pages/Auth.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -7,7 +8,7 @@ const router = createRouter({
     {
       path: '/',
       name: 'entrance',
-      component: RegisterOrLogin,
+      component: Auth,
     },
     {
       path: '/about',
